Optimised Chat Event structure and UI
